requirements.txt
fastapi
psycopg2-binary
uvicorn
mangum
dotenv
config
pydantic_settings
sqlalchemy<2
